requirements/req.txt
numpy==1.17.4
pandas==0.25.3
env==0.1.0
dnspython-1.16.0
requests==2.21.0
googlemaps==3.1.4
pymongo-3.10.0
Flask==1.1.1
Flask-PyMongo==2.3.0
folium-0.10.1
pickle-mixin-1.0.2